removing duplicate gaqs estimation logic
