+++ src/asset-swapper/utils/protocol_fee_utils.ts
@@ -1,4 +1,3 @@
-import { BigNumber } from '@0x/utils';
 import * as heartbeats from 'heartbeats';
 
 import { constants } from '../constants';
@@ -6,11 +5,12 @@ import { SwapQuoterError } from '../types';
 
 const MAX_ERROR_COUNT = 5;
 
-interface GasOracleResponse {
-    result: {
-        // gas price in wei
-        fast: number;
-    };
+interface GasPrices {
+    fast: number;
+    l1CalldataPricePerUnit?: number;
+}
+interface GasInfoResponse {
+    result: GasPrices;
 }
 
 export class ProtocolFeeUtils {
@@ -18,7 +18,7 @@ export class ProtocolFeeUtils {
     private readonly _zeroExGasApiUrl: string;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any -- TODO: fix me!
     private readonly _gasPriceHeart: any;
-    private _gasPriceEstimation: BigNumber = constants.ZERO_AMOUNT;
+    private _gasPriceEstimation: GasPrices | undefined;
     private _errorCount = 0;
 
     public static getInstance(
@@ -42,15 +42,16 @@ export class ProtocolFeeUtils {
     }
 
     /** @returns gas price (in wei) */
-    public async getGasPriceEstimationOrThrowAsync(shouldHardRefresh?: boolean): Promise<BigNumber> {
-        if (this._gasPriceEstimation.eq(constants.ZERO_AMOUNT)) {
-            return this._getGasPriceFromGasStationOrThrowAsync();
+    public async getGasPriceEstimationOrThrowAsync(defaultGasPrices: GasPrices): Promise<GasPrices> {
+        await this._getGasPriceFromGasStationOrThrowAsync();
+        if (this._gasPriceEstimation !== undefined) {
+            return {
+                ...defaultGasPrices,
+                ...this._gasPriceEstimation,
+            };
         }
-        if (shouldHardRefresh) {
-            return this._getGasPriceFromGasStationOrThrowAsync();
-        } else {
-            return this._gasPriceEstimation;
-        }
+
+        return defaultGasPrices;
     }
 
     /**
@@ -66,28 +67,26 @@ export class ProtocolFeeUtils {
         this._initializeHeartBeat();
     }
 
-    private async _getGasPriceFromGasStationOrThrowAsync(): Promise<BigNumber> {
+    private async _getGasPriceFromGasStationOrThrowAsync(): Promise<void> {
         try {
             const res = await fetch(this._zeroExGasApiUrl);
-            const gasInfo: GasOracleResponse = await res.json();
-            const gasPriceWei = new BigNumber(gasInfo.result.fast);
+            const gasInfo: GasInfoResponse = await res.json();
             // Reset the error count to 0 once we have a successful response
             this._errorCount = 0;
-            return gasPriceWei;
+            this._gasPriceEstimation = gasInfo.result;
         } catch (e) {
             this._errorCount++;
             // If we've reached our max error count then throw
-            if (this._errorCount > MAX_ERROR_COUNT || this._gasPriceEstimation.isZero()) {
+            if (this._errorCount > MAX_ERROR_COUNT || this._gasPriceEstimation === undefined) {
                 this._errorCount = 0;
                 throw new Error(SwapQuoterError.NoGasPriceProvidedOrEstimated);
             }
-            return this._gasPriceEstimation;
         }
     }
 
     private _initializeHeartBeat(): void {
         this._gasPriceHeart.createEvent(1, async () => {
-            this._gasPriceEstimation = await this._getGasPriceFromGasStationOrThrowAsync();
+            await this._getGasPriceFromGasStationOrThrowAsync();
         });
     }
 }
