+++ src/handlers/swap_handlers.ts
@@ -13,6 +13,7 @@ import {
     NATIVE_FEE_TOKEN_BY_CHAIN_ID,
     SELL_SOURCE_FILTER_BY_CHAIN_ID,
 } from '../asset-swapper/utils/market_operation_utils/constants';
+import { constants } from '../asset-swapper/constants';
 import {
     CHAIN_ID,
     getIntegratorByIdOrThrow,
@@ -46,7 +47,7 @@ import { parseUtils } from '../utils/parse_utils';
 import { priceComparisonUtils } from '../utils/price_comparison_utils';
 import { publishQuoteReport } from '../utils/quote_report_utils';
 import { schemaUtils } from '../utils/schema_utils';
-import { zeroExGasApiUtils } from '../utils/zero_ex_gas_api_utils';
+import { ProtocolFeeUtils } from '../asset-swapper';
 
 let kafkaProducer: Producer | undefined;
 if (KAFKA_BROKERS !== undefined) {
@@ -300,9 +301,11 @@ export class SwapHandlers {
 
             // Add additional L1 gas cost.
             if (CHAIN_ID === ChainId.Arbitrum) {
-                const gasPrices = await zeroExGasApiUtils.getGasPricesOrDefault({
+                const gasUtils = ProtocolFeeUtils.getInstance(constants.PROTOCOL_FEE_UTILS_POLLING_INTERVAL_IN_MS);
+                const gasPrices = await gasUtils.getGasPriceEstimationOrThrowAsync({
                     fast: 100_000_000, // 0.1 gwei in wei
                 });
+
                 const l1GasCostEstimate = new BigNumber(
                     estimateArbitrumL1CalldataGasCost({
                         l2GasPrice: gasPrices.fast,
